프로필 생성, 수정 로직 합침
