+++ src/repositories/profileMover.respository.ts
@@ -4,10 +4,10 @@
  * 프로필 관련 유저 데이터를 다루는 repository 모듈
  */
 
-import { Mover, MoveType, Prisma } from "@prisma/client";
+import { Mover, Prisma } from "@prisma/client";
 import prisma from "../configs/prisma.config";
 import { MoverProfile } from "../types";
-import { BadRequestError, NotFoundError } from "../types/errors";
+import { BadRequestError } from "../types/errors";
 import { ErrorMessage } from "../constants/ErrorMessage";
 
 //기사님 찾기
@@ -22,45 +22,53 @@ async function findById(id: Mover["id"]) {
   return { ...mover, userType: "mover" };
 }
 
+//지역 라벨 > 지역 아이디 찾기 (관계형이라서 stirng > id로 변환해줘야함)
+//TODO: 지역 레포단을 만들 필요는 없을 것 같아서 우선 프로필 레포에 작성함
+async function findRegionByLabel(user: MoverProfile) {
+  return await prisma.region.findMany({
+    where: {
+      regionName: {
+        in: user.serviceArea,
+      },
+    },
+  });
+}
+
 //기사님 프로필 생성/수정
-async function modifyMoverProfile(user: MoverProfile) {
+async function modifyMoverProfile(user: MoverProfile, updateData: Prisma.MoverUpdateInput) {
   try {
-    //업데이트할 데이터 목록
-    const updateData: Prisma.MoverUpdateInput = {
-      profileImage: user.image,
-      nickName: user.nickName,
-      career: user.career,
-      introduction: user.introduction,
-      description: user.description,
-      isProfileCompleted: true,
-    };
+    //TODO: 삭제 예정 //업데이트할 데이터 목록
+    // const updateData: Prisma.MoverUpdateInput = {
+    //   profileImage: user.image,
+    //   nickName: user.nickName,
+    //   career: user.career,
+    //   introduction: user.introduction,
+    //   description: user.description,
+    //   isProfileCompleted: true,
+    // };
 
-    //서비스 종류
-    if (user.serviceType && user.serviceType.length > 0) {
-      updateData.serviceType = {
-        set: user.serviceType as MoveType[], // (enum 타입)
-      };
-    }
+    // //서비스 종류
+    // if (user.serviceType && user.serviceType.length > 0) {
+    //   updateData.serviceType = {
+    //     set: user.serviceType as MoveType[], // (enum 타입)
+    //   };
+    // }
 
-    //서비스 지역
-    if (user.serviceArea && user.serviceArea.length > 0) {
-      const matchedRegions = await prisma.region.findMany({
-        // (관계형이라서 stirng > id로 변환해줘야함)
-        where: {
-          regionName: {
-            in: user.serviceArea,
-          },
-        },
-      });
+    // //서비스 지역
+    // if (user.serviceArea && user.serviceArea.length > 0) {
+    //   const matchedRegions = await prisma.region.findMany({
+    //     // (관계형이라서 stirng > id로 변환해줘야함)
+    //     where: {
+    //       regionName: {
+    //         in: user.serviceArea,
+    //       },
+    //     },
+    //   });
 
-      if (matchedRegions.length !== user.serviceArea.length) {
-        throw new BadRequestError(ErrorMessage.REGION_NOT_FOUND); // 없는 지역이 경우 에러
-      }
-
-      updateData.serviceArea = {
-        set: matchedRegions.map((region) => ({ id: region.id })),
-      };
-    }
+    //   updateData.serviceArea = {
+    //     set: matchedRegions.map((region) => ({ id: region.id })),
+    //   };
+    // }
 
     const modifiedMoverProfile = await prisma.mover.update({
       where: { id: user.userId },
@@ -78,5 +86,6 @@ async function modifyMoverProfile(user: MoverProfile) {
 
 export default {
   findById,
+  findRegionByLabel,
   modifyMoverProfile,
 };
