+++ src/routers/profile.router.ts
@@ -2,7 +2,7 @@ import express from "express";
 import profileClientController from "../controllers/profileClient.controller";
 import { validateReq, verifyAccessToken } from "../middlewares/auth.middleware";
 import profileMoverController from "../controllers/profileMover.controller";
-import { MoverProfileSchema } from "../dtos/profileClient.dto";
+import { MoverProfileSchema } from "../dtos/mover.dto";
 
 const profileRouter = express.Router();
 
@@ -11,15 +11,15 @@ const profileRouter = express.Router();
  * @description 프로필 관련 라우트 정의 모듈 (프로필 등록, 프로필 수정 등)
  */
 
-//기사님 프로필 등록
-profileRouter.post(
-  "/mover",
-  verifyAccessToken,
-  validateReq(MoverProfileSchema),
-  profileMoverController.moverCreateProfile,
-);
+//  TODO 삭제 예정//기사님 프로필 등록
+// profileRouter.post(
+//   "/mover",
+//   verifyAccessToken,
+//   validateReq(MoverProfileSchema),
+//   profileMoverController.moverCreateProfile,
+// );
 
-//기사님 프로필 수정
+//기사님 프로필 생성&수정
 profileRouter.patch(
   "/mover",
   verifyAccessToken,
