+++ src/types/index.ts
@@ -8,8 +8,8 @@
  * export * from "./service";
  */
 
-export * from "./authMover.type";
+export * from "./auth.types";
 export * from "./token.type";
-export * from "./authClient.types";
+export * from "./auth.types";
 export * from "./profile.types";
 export * from "./request.type";
